+++ src/SmartMap.php
@@ -54,7 +54,7 @@ class SmartMap extends Plugin
     public $hasCpSettings = true;
 
     /** @var bool $schemaVersion Current schema version of the plugin. */
-    public $schemaVersion = '3.2.2';
+    public $schemaVersion = '3.2.3';
 
     /** @inheritDoc */
     public function init()