+++ CHANGELOG.md
@@ -5,6 +5,9 @@
 ### Changed
 - Craft 3.1 is now required.
 
+### Fixed
+- Replaced problematic migration.
+
 ## 3.2.2.1 - 2019-04-08
 
 ### Changed