+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - KML files can now be specified by absolute URL.
 - Added [`field`](https://www.doublesecretagency.com/plugins/smart-map/docs/customizing-the-map-in-twig) option to filter field(s) to be included on map.
 - Added [`hasCoords`](https://www.doublesecretagency.com/plugins/smart-map/docs/filtering-out-entries-with-invalid-coordinates) parameter to filter out results with incomplete coordinates.
+- All API access keys (for Google, ipstack, and MaxMind) can now be set via [environment variables](https://docs.craftcms.com/v3/config/environments.html).
 
 ### Changed
 - Craft 3.1 is now required.
@@ -17,6 +18,7 @@
 - Replaced problematic migration.
 - Fixed PHP 7.4 compatibility issues.
 - Fixed error when passing icon array to static map.
+- Fixed all bugs in the debug page.
 
 ## 3.2.2.1 - 2019-04-08